slight restructure, removed PInvoke and replaced with Native. Removed no longer necessary pinvoke calls
