+++ Payload_Type/athena/athena/agent_code/Agent.Managers.Linux/Native.cs
@@ -7,7 +7,7 @@ using Microsoft.Win32.SafeHandles;
 
 namespace Agent.Managers
 {
-    public static class PInvoke
+    public static class Native
     {
         [DllImport("libc")]
         public static extern uint geteuid();
