+++ Payload_Type/athena/athena/agent_code/Agent.Managers.Windows/Native.cs
@@ -1,14 +1,14 @@
 ﻿using Microsoft.Win32.SafeHandles;
 using System.Runtime.InteropServices;
 
-namespace Agent.Utilities.Invoker
+namespace Agent.Utilities
 {
     public class Native
     {
         [DllImport("kernel32.dll")]
         [return: MarshalAs(UnmanagedType.Bool)]
         public static extern bool CreateProcess(
-        string lpApplicationName,
+           string lpApplicationName,
            string lpCommandLine,
            SECURITY_ATTRIBUTES lpProcessAttributes,
            SECURITY_ATTRIBUTES lpThreadAttributes,
@@ -31,6 +31,7 @@ namespace Agent.Utilities.Invoker
         public static extern bool PeekNamedPipe(IntPtr handle,
             IntPtr buffer, IntPtr nBufferSize, IntPtr bytesRead,
             ref uint bytesAvail, IntPtr BytesLeftThisMessage);
+        
         [DllImport("kernel32.dll", SetLastError = true)]
         [return: MarshalAs(UnmanagedType.Bool)]
         public static extern bool UpdateProcThreadAttribute(
@@ -51,11 +52,6 @@ namespace Agent.Utilities.Invoker
 
         [DllImport("kernel32.dll", SetLastError = true)]
         public static extern bool CloseHandle(IntPtr hObject);
-
-
-        [DllImport("kernel32.dll", SetLastError = true)]
-        public static extern int GetCurrentThread();
-
 
         [DllImport("kernel32.dll", SetLastError = true)]
         public static extern IntPtr OpenProcess(
@@ -80,15 +76,6 @@ namespace Agent.Utilities.Invoker
             ref UInt32 returnLength
         );
 
-        [DllImport("kernel32.dll", SetLastError = true, ExactSpelling = true)]
-        public static extern IntPtr VirtualAllocEx(IntPtr hProcess, IntPtr lpAddress, uint dwSize, AllocationType flAllocationType, MemoryProtection flProtect);
-
-        [DllImport("kernel32.dll")]
-        public static extern bool WriteProcessMemory(IntPtr hProcess, IntPtr lpBaseAddress, byte[] lpBuffer, int nSize, out IntPtr lpNumberOfBytesWritten);
-
-        [DllImport("kernel32.dll")]
-        public static extern IntPtr CreateRemoteThread(IntPtr hProcess, IntPtr lpThreadAttributes, uint dwStackSize, IntPtr lpStartAddress, IntPtr lpParameter, ThreadCreationFlags dwCreationFlags, out IntPtr lpThreadId);
-
         [DllImport("ntdll.dll", SetLastError = true)]
         public static extern Boolean NtReadVirtualMemory(
             IntPtr ProcessHandle,
@@ -97,21 +84,6 @@ namespace Agent.Utilities.Invoker
             UInt32 NumberOfBytesToRead,
             ref UInt32 liRet
         );
-
-        [DllImport("ntdll.dll", SetLastError = true)]
-        public static extern IntPtr RtlCreateUserThread(IntPtr processHandle, IntPtr threadSecurity, bool createSuspended, Int32 stackZeroBits, IntPtr stackReserved, IntPtr stackCommit, IntPtr startAddress, IntPtr parameter, ref IntPtr threadHandle, CLIENT_ID clientId);
-
-        [DllImport("ntdll.dll", SetLastError = true)]
-        public static extern UInt32 ZwQueryInformationProcess(IntPtr hProcess, int procInformationClass, ref PROCESS_BASIC_INFORMATION procInformation, UInt32 ProcInfoLen, ref UInt32 retlen);
-
-        [DllImport("kernel32.dll", SetLastError = true)]
-        public static extern bool ReadProcessMemory(IntPtr hProcess, IntPtr lpBaseAddress, byte[] lpBuffer, int dwSize, out IntPtr lpNumberOfBytesRead);
-
-        [DllImport("ntdll.dll", SetLastError = true, ExactSpelling = true)]
-        public static extern UInt32 NtCreateSection(ref IntPtr SectionHandle, SectionAccess DesiredAccess, IntPtr ObjectAttributes, ref UInt64 MaximumSize, MemoryProtection SectionPageProtection, MappingAttributes AllocationAttributes, IntPtr FileHandle);
-
-        [DllImport("ntdll.dll", SetLastError = true)]
-        public static extern UInt32 NtMapViewOfSection(IntPtr SectionHandle, IntPtr ProcessHandle, ref IntPtr BaseAddress, UIntPtr ZeroBits, UIntPtr CommitSize, ref UInt64 SectionOffset, ref UInt64 ViewSize, uint InheritDisposition, UInt32 AllocationType, MemoryProtection Win32Protect);
 
         [DllImport("kernel32.dll")]
         public static extern bool CreatePipe(out IntPtr hReadPipe, out IntPtr hWritePipe,
@@ -128,168 +100,7 @@ namespace Agent.Utilities.Invoker
 
         [DllImport("kernel32.dll", SetLastError = true)]
         public static extern uint ResumeThread(IntPtr hThread);
-        public enum NTSTATUS : uint
-        {
-            Success = 0,
-            Informational = 0x40000000,
-            Error = 0xc0000000
-        }
 
-        [StructLayout(LayoutKind.Explicit, Size = 18)]
-        public struct CURDIR
-        {
-            [FieldOffset(0)]
-            public UNICODE_STRING DosPath;
-            [FieldOffset(16)]
-            public IntPtr Handle;
-        }
-
-        [StructLayout(LayoutKind.Sequential, Pack = 0)]
-        public struct CLIENT_ID
-        {
-            public IntPtr UniqueProcess;
-            public IntPtr UniqueThread;
-        }
-
-        [Flags]
-        public enum ProcessAccessFlags : uint
-        {
-            All = 0x001F0FFF,
-            Terminate = 0x00000001,
-            CreateThread = 0x00000002,
-            VirtualMemoryOperation = 0x00000008,
-            VirtualMemoryRead = 0x00000010,
-            VirtualMemoryWrite = 0x00000020,
-            DuplicateHandle = 0x00000040,
-            CreateProcess = 0x000000080,
-            SetQuota = 0x00000100,
-            SetInformation = 0x00000200,
-            QueryInformation = 0x00000400,
-            QueryLimitedInformation = 0x00001000,
-            Synchronize = 0x00100000
-        }
-
-        [Flags]
-        public enum SectionAccess : UInt32
-        {
-            SECTION_EXTEND_SIZE = 0x0010,
-            SECTION_QUERY = 0x0001,
-            SECTION_MAP_WRITE = 0x0002,
-            SECTION_MAP_READ = 0x0004,
-            SECTION_MAP_EXECUTE = 0x0008,
-            SECTION_ALL_ACCESS = 0xe
-        }
-
-
-        [Flags]
-        public enum ProcessParametersFlags : uint
-        {
-            NORMALIZED = 0x01,
-            PROFILE_USER = 0x02,
-            PROFILE_SERVER = 0x04,
-            PROFILE_KERNEL = 0x08,
-            UNKNOWN = 0x10,
-            RESERVE_1MB = 0x20,
-            DISABLE_HEAP_CHECKS = 0x100,
-            PROCESS_OR_1 = 0x200,
-            PROCESS_OR_2 = 0x400,
-            PRIVATE_DLL_PATH = 0x1000,
-            LOCAL_DLL_PATH = 0x2000,
-            NX = 0x20000,
-        }
-
-
-        [Flags]
-        public enum CreateProcessFlags
-        {
-            CREATE_BREAKAWAY_FROM_JOB = 0x01000000,
-            CREATE_DEFAULT_ERROR_MODE = 0x04000000,
-            CREATE_NEW_CONSOLE = 0x00000010,
-            CREATE_NEW_PROCESS_GROUP = 0x00000200,
-            CREATE_NO_WINDOW = 0x08000000,
-            CREATE_PROTECTED_PROCESS = 0x00040000,
-            CREATE_PRESERVE_CODE_AUTHZ_LEVEL = 0x02000000,
-            CREATE_SEPARATE_WOW_VDM = 0x00000800,
-            CREATE_SHARED_WOW_VDM = 0x00001000,
-            CREATE_SUSPENDED = 0x00000004,
-            CREATE_UNICODE_ENVIRONMENT = 0x00000400,
-            DEBUG_ONLY_THIS_PROCESS = 0x00000002,
-            DEBUG_PROCESS = 0x00000001,
-            DETACHED_PROCESS = 0x00000008,
-            EXTENDED_STARTUPINFO_PRESENT = 0x00080000,
-            INHERIT_PARENT_AFFINITY = 0x00010000
-        }
-
-        [Flags]
-        public enum MappingAttributes : UInt32
-        {
-            SEC_COMMIT = 0x8000000,
-            SEC_IMAGE = 0x1000000,
-            SEC_IMAGE_NO_EXECUTE = 0x11000000,
-            SEC_LARGE_PAGES = 0x80000000,
-            SEC_NOCACHE = 0x10000000,
-            SEC_RESERVE = 0x4000000,
-            SEC_WRITECOMBINE = 0x40000000
-        }
-
-        [Flags]
-        public enum AllocationType
-        {
-            NULL = 0x0,
-            Commit = 0x1000,
-            Reserve = 0x2000,
-            Decommit = 0x4000,
-            Release = 0x8000,
-            Reset = 0x80000,
-            Physical = 0x400000,
-            TopDown = 0x100000,
-            WriteWatch = 0x200000,
-            LargePages = 0x20000000
-        }
-        public enum MemoryProtection : uint
-        {
-            PAGE_EXECUTE = 0x00000010,
-            PAGE_EXECUTE_READ = 0x00000020,
-            PAGE_EXECUTE_READWRITE = 0x00000040,
-            PAGE_EXECUTE_WRITECOPY = 0x00000080,
-            PAGE_NOACCESS = 0x00000001,
-            PAGE_READONLY = 0x00000002,
-            PAGE_READWRITE = 0x00000004,
-            PAGE_WRITECOPY = 0x00000008,
-            PAGE_GUARD = 0x00000100,
-            PAGE_NOCACHE = 0x00000200,
-            PAGE_WRITECOMBINE = 0x00000400
-        }
-        public enum ThreadCreationFlags : uint
-        {
-            NORMAL = 0x0,
-            CREATE_SUSPENDED = 0x00000004,
-            STACK_SIZE_PARAM_IS_A_RESERVATION = 0x00010000
-        }
-
-
-
-        [StructLayout(LayoutKind.Sequential)]
-        public struct PROCESS_BASIC_INFORMATION
-        {
-            public IntPtr ExitStatus;
-            public IntPtr PebBaseAddress;
-            public IntPtr AffinityMask;
-            public IntPtr BasePriority;
-            public UIntPtr UniqueProcessId;
-            public IntPtr InheritedFromUniqueProcessId;
-        }
-
-        /*
-        [StructLayout(LayoutKind.Sequential)]
-        public struct _RTL_DRIVE_LETTER_CURDIR
-        {
-            UInt16 Flags;
-            UInt16 Length;
-            UInt32 TimeStamp;
-            UNICODE_STRING DosPath;
-        }
-        */
         [DllImport("kernel32.dll", SetLastError = true)]
         public static extern IntPtr OpenProcess(ProcessAccessFlags processAccess, bool bInheritHandle, int processId);
 
@@ -299,14 +110,31 @@ namespace Agent.Utilities.Invoker
         [DllImport("kernel32.dll", SetLastError = true)]
         [return: MarshalAs(UnmanagedType.Bool)]
         public static extern bool DuplicateHandle(IntPtr hSourceProcessHandle,
-           SafeFileHandle hSourceHandle, IntPtr hTargetProcessHandle, ref SafeFileHandle lpTargetHandle,
+           IntPtr hSourceHandle, IntPtr hTargetProcessHandle, ref IntPtr lpTargetHandle,
            uint dwDesiredAccess, [MarshalAs(UnmanagedType.Bool)] bool bInheritHandle, uint dwOptions);
 
-        [DllImport("kernel32.dll", SetLastError = true)]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        public static extern bool DuplicateHandle(IntPtr hSourceProcessHandle,
-   IntPtr hSourceHandle, IntPtr hTargetProcessHandle, ref IntPtr lpTargetHandle,
-   uint dwDesiredAccess, [MarshalAs(UnmanagedType.Bool)] bool bInheritHandle, uint dwOptions);
+        [DllImport("advapi32.dll", CharSet = CharSet.Unicode, SetLastError = true)]
+        public static extern Boolean LogonUser(
+            String lpszUserName,
+            String lpszDomain,
+            String lpszPassword,
+            LogonType dwLogonType,
+            LogonProvider dwLogonProvider,
+            out SafeAccessTokenHandle phToken);
+
+
+        [DllImport("advapi32.dll", SetLastError = true)]
+        public static extern bool ImpersonateLoggedOnUser(SafeAccessTokenHandle hToken);
+
+        [DllImport("advapi32.dll", SetLastError = true)]
+        public static extern bool RevertToSelf();
+
+        [DllImport("kernel32.dll")]
+        public static extern IntPtr GetConsoleWindow();
+
+        [DllImport("user32.dll")]
+        public static extern bool ShowWindow(IntPtr hWnd, int nCmdShow);
+
 
         [StructLayout(LayoutKind.Explicit, Size = 136)]
         public struct RTL_USER_PROCESS_PARAMETERS
@@ -326,13 +154,6 @@ namespace Agent.Utilities.Invoker
                                        //[MarshalAs(UnmanagedType.ByValArray, SizeConst = 0x20)]
                                        //public UNICODE_STRING DLCurrentDirectory;
         };
-
-        [StructLayout(LayoutKind.Explicit, Size = 8)]
-        struct LARGE_INTEGER
-        {
-            [FieldOffset(0)] public UInt32 LowPart;
-            [FieldOffset(4)] public Int32 HighPart;
-        }
 
         [StructLayout(LayoutKind.Explicit, Size = 64)]
         public struct PEB
@@ -380,7 +201,6 @@ namespace Agent.Utilities.Invoker
             }
         }
 
-
         [StructLayout(LayoutKind.Sequential)]
         public struct SECURITY_ATTRIBUTES
         {
@@ -389,14 +209,12 @@ namespace Agent.Utilities.Invoker
             public bool bInheritHandle;
         }
 
-
         [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
         public struct STARTUPINFOEX
         {
             public STARTUPINFO StartupInfo;
             public IntPtr lpAttributeList;
         }
-
 
         [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
         public struct STARTUPINFO
@@ -421,7 +239,6 @@ namespace Agent.Utilities.Invoker
             public IntPtr hStdError;
         }
 
-
         [StructLayout(LayoutKind.Sequential)]
         public struct PROCESS_INFORMATION
         {
@@ -442,5 +259,89 @@ namespace Agent.Utilities.Invoker
         public const int STARTF_USESTDHANDLES = 0x00000100;
         public const int STARTF_USESHOWWINDOW = 0x00000001;
         public const short SW_HIDE = 0x0000;
+
+        [Flags]
+        public enum LogonFlags
+        {
+            LOGON_WITH_PROFILE = 1,
+            LOGON_NETCREDENTIALS_ONLY = 2
+        }
+
+        [Flags]
+        public enum LogonType
+        {
+            LOGON32_LOGON_INTERACTIVE = 2,
+            LOGON32_LOGON_NETWORK = 3,
+            LOGON32_LOGON_BATCH = 4,
+            LOGON32_LOGON_SERVICE = 5,
+            LOGON32_LOGON_UNLOCK = 7,
+            LOGON32_LOGON_NETWORK_CLEARTEXT = 8,
+            LOGON32_LOGON_NEW_CREDENTIALS = 9
+        }
+
+        [Flags]
+        public enum LogonProvider
+        {
+            LOGON32_PROVIDER_DEFAULT = 0,
+            LOGON32_PROVIDER_WINNT35,
+            LOGON32_PROVIDER_WINNT40,
+            LOGON32_PROVIDER_WINNT50
+        }
+        public enum NTSTATUS : uint
+        {
+            Success = 0,
+            Informational = 0x40000000,
+            Error = 0xc0000000
+        }
+
+        [Flags]
+        public enum ProcessAccessFlags : uint
+        {
+            All = 0x001F0FFF,
+            Terminate = 0x00000001,
+            CreateThread = 0x00000002,
+            VirtualMemoryOperation = 0x00000008,
+            VirtualMemoryRead = 0x00000010,
+            VirtualMemoryWrite = 0x00000020,
+            DuplicateHandle = 0x00000040,
+            CreateProcess = 0x000000080,
+            SetQuota = 0x00000100,
+            SetInformation = 0x00000200,
+            QueryInformation = 0x00000400,
+            QueryLimitedInformation = 0x00001000,
+            Synchronize = 0x00100000
+        }
+
+        [Flags]
+        public enum CreateProcessFlags
+        {
+            CREATE_BREAKAWAY_FROM_JOB = 0x01000000,
+            CREATE_DEFAULT_ERROR_MODE = 0x04000000,
+            CREATE_NEW_CONSOLE = 0x00000010,
+            CREATE_NEW_PROCESS_GROUP = 0x00000200,
+            CREATE_NO_WINDOW = 0x08000000,
+            CREATE_PROTECTED_PROCESS = 0x00040000,
+            CREATE_PRESERVE_CODE_AUTHZ_LEVEL = 0x02000000,
+            CREATE_SEPARATE_WOW_VDM = 0x00000800,
+            CREATE_SHARED_WOW_VDM = 0x00001000,
+            CREATE_SUSPENDED = 0x00000004,
+            CREATE_UNICODE_ENVIRONMENT = 0x00000400,
+            DEBUG_ONLY_THIS_PROCESS = 0x00000002,
+            DEBUG_PROCESS = 0x00000001,
+            DETACHED_PROCESS = 0x00000008,
+            EXTENDED_STARTUPINFO_PRESENT = 0x00080000,
+            INHERIT_PARENT_AFFINITY = 0x00010000
+        }
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct PROCESS_BASIC_INFORMATION
+        {
+            public IntPtr ExitStatus;
+            public IntPtr PebBaseAddress;
+            public IntPtr AffinityMask;
+            public IntPtr BasePriority;
+            public UIntPtr UniqueProcessId;
+            public IntPtr InheritedFromUniqueProcessId;
+        }
     }
 }
