+++ Payload_Type/athena/athena/agent_code/Agent.Managers.Windows/ProcessSpawner.cs
@@ -3,10 +3,8 @@ using Microsoft.Win32.SafeHandles;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
 using System.Text;
-using Agent.Utilities.Invoker;
+using Agent.Utilities;
 using Agent.Models;
-using System.Linq.Expressions;
-using System.Net;
 
 namespace Agent.Utlities
 {
