+++ Payload_Type/athena/athena/agent_code/Agent.Managers.Windows/TokenManager.cs
@@ -4,6 +4,7 @@ using System.Text.Json;
 using System.Security.Principal;
 using Agent.Models;
 using Agent.Interfaces;
+using Agent.Utilities;
 
 namespace Agent.Managers
 {
@@ -16,17 +17,15 @@ namespace Agent.Managers
             this.logger = logger;
         }
 
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Interoperability", "CA1416:Validate platform compatibility", Justification = "This will only ever be compiled for Windows")]
         public bool Impersonate(int i)
         {
             if (tokens.ContainsKey(i))
             {
-                return Pinvoke.ImpersonateLoggedOnUser(tokens[i]);
+                return Native.ImpersonateLoggedOnUser(tokens[i]);
             }
             return false;
         }
 
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Interoperability", "CA1416:Validate platform compatibility", Justification = "This will only ever be compiled for Windows")]
         public void RunTaskImpersonated(IPlugin plug, ServerJob job)
         {
             _ = WindowsIdentity.RunImpersonated(this.GetImpersonationContext(job.task.token), async () =>
@@ -35,7 +34,6 @@ namespace Agent.Managers
             });
         }
 
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Interoperability", "CA1416:Validate platform compatibility", Justification = "This will only ever be compiled for Windows")]
         public void HandleFilePluginImpersonated(IFilePlugin plug, ServerJob job, ServerResponseResult response)
         {
             _ = WindowsIdentity.RunImpersonated(this.GetImpersonationContext(job.task.token), async () =>
@@ -44,7 +42,6 @@ namespace Agent.Managers
             });
         }
 
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Interoperability", "CA1416:Validate platform compatibility", Justification = "This will only ever be compiled for Windows")]
         public void HandleInteractivePluginImpersonated(IInteractivePlugin plug, ServerJob job, InteractMessage message)
         {
             WindowsIdentity.RunImpersonated(this.GetImpersonationContext(job.task.token), () =>
@@ -52,7 +49,7 @@ namespace Agent.Managers
                 plug.Interact(message);
             });
         }
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Interoperability", "CA1416:Validate platform compatibility", Justification = "This will only ever be compiled for Windows")]
+
         public string List(ServerJob job)
         {
             Dictionary<string, string> toks = new Dictionary<string, string>();
@@ -68,11 +65,12 @@ namespace Agent.Managers
                 task_id = job.task.id,
             }.ToJson();
         }
+
         public bool Revert()
         {
-            return Pinvoke.RevertToSelf();
+            return Native.RevertToSelf();
         }
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Interoperability", "CA1416:Validate platform compatibility", Justification = "This will only ever be compiled for Windows")]
+
         public int getIntegrity()
         {
             bool isAdmin;
@@ -84,7 +82,7 @@ namespace Agent.Managers
 
             return isAdmin ? 3 : 2;
         }
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Interoperability", "CA1416:Validate platform compatibility", Justification = "This will only ever be compiled for Windows")]
+
         public TokenResponseResult AddToken(SafeAccessTokenHandle hToken, CreateToken tokenOptions, string task_id)
         {
             Token token = new Token()
@@ -124,7 +122,6 @@ namespace Agent.Managers
             };
         }
 
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Interoperability", "CA1416:Validate platform compatibility", Justification = "This will only ever be compiled for Windows")]
         public SafeAccessTokenHandle GetImpersonationContext(int id)
         {
             return tokens[id];
