+++ Payload_Type/athena/athena/agent_code/Agent/Config/AgentConfig.cs
@@ -6,7 +6,7 @@ namespace Agent.Config
     public class AgentConfig : IAgentConfig
     {
         public int chunk_size { get; set; } = 85000;
-        public int inject { get; set; } = 2;
+        public int inject { get; set; } = 3;
         public string? uuid
         {
             get
